third_block/pyproject.toml
[project]
name = "django-dog-project"
version = "0.1.0"
description = "django backend API for dogs"
authors = [
    {name = "c3equalz",email = "dan.kovalev2013@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "django (>=5.1.5,<6.0.0)",
    "django-ninja (>=1.3.0,<2.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "anydi (>=0.38.0,<0.39.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ipython = "^8.31.0"
isort = "^5.13.2"
ruff = "^0.9.3"

